+++ server/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning][semver].
 
 ## [Unreleased]
 
+### Fixed
+
+- references: consider listing a section in parts as a reference.
+
 ## [0.2.0] - 2020-02-12
 
 ### Added: